@@ -1,29 +1,26 @@
 """Anomalib dataset base class."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-
-from __future__ import annotations
 
 import copy
 import logging
-from abc import ABC, abstractmethod
+from abc import ABC
+from collections.abc import Sequence
 from pathlib import Path
-from typing import Sequence
 
-import albumentations as A
-import cv2
-import numpy as np
 import pandas as pd
+import torch
 from pandas import DataFrame
-from torch import Tensor
 from torch.utils.data import Dataset
+from torchvision.transforms.v2 import Transform
+from torchvision.tv_tensors import Mask
 
-from anomalib.data.task_type import TaskType
-from anomalib.data.utils import masks_to_boxes, read_image
+from anomalib import TaskType
+from anomalib.data.utils import LabelName, masks_to_boxes, read_image, read_mask
 
 _EXPECTED_COLUMNS_CLASSIFICATION = ["image_path", "split"]
-_EXPECTED_COLUMNS_SEGMENTATION = _EXPECTED_COLUMNS_CLASSIFICATION + ["mask_path"]
+_EXPECTED_COLUMNS_SEGMENTATION = [*_EXPECTED_COLUMNS_CLASSIFICATION, "mask_path"]
 _EXPECTED_COLUMNS_PERTASK = {
     "classification": _EXPECTED_COLUMNS_CLASSIFICATION,
     "segmentation": _EXPECTED_COLUMNS_SEGMENTATION,
@@ -36,43 +33,80 @@ logger = logging.getLogger(__name__)
 class AnomalibDataset(Dataset, ABC):
     """Anomalib dataset.
 
+    The dataset is based on a dataframe that contains the information needed by the dataloader to load each of
+    the dataset items into memory.
+
+    The samples dataframe must be set from the subclass using the setter of the `samples` property.
+
+    The DataFrame must, at least, include the following columns:
+        - `split` (str): The subset to which the dataset item is assigned (e.g., 'train', 'test').
+        - `image_path` (str): Path to the file system location where the image is stored.
+        - `label_index` (int): Index of the anomaly label, typically 0 for 'normal' and 1 for 'anomalous'.
+        - `mask_path` (str, optional): Path to the ground truth masks (for the anomalous images only).
+        Required if task is 'segmentation'.
+
+    Example DataFrame:
+        +---+-------------------+-----------+-------------+------------------+-------+
+        |   | image_path        | label     | label_index | mask_path        | split |
+        +---+-------------------+-----------+-------------+------------------+-------+
+        | 0 | path/to/image.png | anomalous | 1           | path/to/mask.png | train |
+        +---+-------------------+-----------+-------------+------------------+-------+
+
+    Note:
+        The example above is illustrative and may need to be adjusted based on the specific dataset structure.
+
     Args:
         task (str): Task type, either 'classification' or 'segmentation'
-        transform (A.Compose): Albumentations Compose object describing the transforms that are applied to the inputs.
+        transform (Transform, optional): Transforms that should be applied to the input images.
+            Defaults to ``None``.
     """
 
-    def __init__(self, task: TaskType, transform: A.Compose) -> None:
+    def __init__(self, task: TaskType | str, transform: Transform | None = None) -> None:
         super().__init__()
-        self.task = task
+        self.task = TaskType(task)
         self.transform = transform
-        self._samples: DataFrame
+        self._samples: DataFrame | None = None
+        self._category: str | None = None
+
+    @property
+    def name(self) -> str:
+        """Name of the dataset."""
+        class_name = self.__class__.__name__
+
+        # Remove the `_dataset` suffix from the class name
+        if class_name.endswith("Dataset"):
+            class_name = class_name[:-7]
+
+        return class_name
 
     def __len__(self) -> int:
         """Get length of the dataset."""
         return len(self.samples)
 
-    def subsample(self, indices: Sequence[int], inplace: bool = False) -> AnomalibDataset:
+    def subsample(self, indices: Sequence[int], inplace: bool = False) -> "AnomalibDataset":
         """Subsamples the dataset at the provided indices.
 
         Args:
             indices (Sequence[int]): Indices at which the dataset is to be subsampled.
             inplace (bool): When true, the subsampling will be performed on the instance itself.
+                Defaults to ``False``.
         """
-        assert len(set(indices)) == len(indices), "No duplicates allowed in indices."
+        if len(set(indices)) != len(indices):
+            msg = "No duplicates allowed in indices."
+            raise ValueError(msg)
         dataset = self if inplace else copy.deepcopy(self)
         dataset.samples = self.samples.iloc[indices].reset_index(drop=True)
         return dataset
 
     @property
-    def is_setup(self) -> bool:
-        """Checks if setup() been called."""
-        return hasattr(self, "_samples")
-
-    @property
     def samples(self) -> DataFrame:
         """Get the samples dataframe."""
-        if not self.is_setup:
-            raise RuntimeError("Dataset is not setup yet. Call setup() first.")
+        if self._samples is None:
+            msg = (
+                "Dataset does not have a samples dataframe. Ensure that a dataframe has been assigned to "
+                "`dataset.samples`."
+            )
+            raise RuntimeError(msg)
         return self._samples
 
     @samples.setter
@@ -83,102 +117,92 @@ class AnomalibDataset(Dataset, ABC):
             samples (DataFrame): DataFrame with new samples.
         """
         # validate the passed samples by checking the
-        assert isinstance(samples, DataFrame), f"samples must be a pandas.DataFrame, found {type(samples)}"
+        if not isinstance(samples, DataFrame):
+            msg = f"samples must be a pandas.DataFrame, found {type(samples)}"
+            raise TypeError(msg)
+
         expected_columns = _EXPECTED_COLUMNS_PERTASK[self.task]
-        assert all(
-            col in samples.columns for col in expected_columns
-        ), f"samples must have (at least) columns {expected_columns}, found {samples.columns}"
-        assert samples["image_path"].apply(lambda p: Path(p).exists()).all(), "missing file path(s) in samples"
+        if not all(col in samples.columns for col in expected_columns):
+            msg = f"samples must have (at least) columns {expected_columns}, found {samples.columns}"
+            raise ValueError(msg)
+
+        if not samples["image_path"].apply(lambda p: Path(p).exists()).all():
+            msg = "missing file path(s) in samples"
+            raise FileNotFoundError(msg)
 
         self._samples = samples.sort_values(by="image_path", ignore_index=True)
 
     @property
+    def category(self) -> str | None:
+        """Get the category of the dataset."""
+        return self._category
+
+    @category.setter
+    def category(self, category: str) -> None:
+        """Set the category of the dataset."""
+        self._category = category
+
+    @property
     def has_normal(self) -> bool:
         """Check if the dataset contains any normal samples."""
-        return 0 in list(self.samples.label_index)
+        return LabelName.NORMAL in list(self.samples.label_index)
 
     @property
     def has_anomalous(self) -> bool:
         """Check if the dataset contains any anomalous samples."""
-        return 1 in list(self.samples.label_index)
+        return LabelName.ABNORMAL in list(self.samples.label_index)
 
-    def __getitem__(self, index: int) -> dict[str, str | Tensor]:
+    def __getitem__(self, index: int) -> dict[str, str | torch.Tensor]:
         """Get dataset item for the index ``index``.
 
         Args:
             index (int): Index to get the item.
 
         Returns:
-            Union[dict[str, Tensor], dict[str, str | Tensor]]: Dict of image tensor during training.
-                Otherwise, Dict containing image path, target path, image tensor, label and transformed bounding box.
+            dict[str, str | torch.Tensor]: Dict of image tensor during training. Otherwise, Dict containing image path,
+                target path, image tensor, label and transformed bounding box.
         """
+        image_path = self.samples.iloc[index].image_path
+        mask_path = self.samples.iloc[index].mask_path
+        label_index = self.samples.iloc[index].label_index
 
-        image_path = self._samples.iloc[index].image_path
-        mask_path = self._samples.iloc[index].mask_path
-        label_index = self._samples.iloc[index].label_index
-
-        image = read_image(image_path)
-        item = dict(image_path=image_path, label=label_index)
+        image = read_image(image_path, as_tensor=True)
+        item = {"image_path": image_path, "label": label_index}
 
         if self.task == TaskType.CLASSIFICATION:
-            transformed = self.transform(image=image)
-            item["image"] = transformed["image"]
-        elif self.task in (TaskType.DETECTION, TaskType.SEGMENTATION):
+            item["image"] = self.transform(image) if self.transform else image
+        elif self.task in {TaskType.DETECTION, TaskType.SEGMENTATION}:
             # Only Anomalous (1) images have masks in anomaly datasets
             # Therefore, create empty mask for Normal (0) images.
-
-            if label_index == 0:
-                mask = np.zeros(shape=image.shape[:2])
-            else:
-                mask = cv2.imread(mask_path, flags=0) / 255.0
-
-            transformed = self.transform(image=image, mask=mask)
-
-            item["image"] = transformed["image"]
-            item["mask_path"] = mask_path
-            item["mask"] = transformed["mask"]
+            mask = (
+                Mask(torch.zeros(image.shape[-2:])).to(torch.uint8)
+                if label_index == LabelName.NORMAL
+                else read_mask(mask_path, as_tensor=True)
+            )
+            item["image"], item["mask"] = self.transform(image, mask) if self.transform else (image, mask)
 
             if self.task == TaskType.DETECTION:
                 # create boxes from masks for detection task
                 boxes, _ = masks_to_boxes(item["mask"])
                 item["boxes"] = boxes[0]
         else:
-            raise ValueError(f"Unknown task type: {self.task}")
+            msg = f"Unknown task type: {self.task}"
+            raise ValueError(msg)
 
         return item
 
-    def __add__(self, other_dataset: AnomalibDataset) -> AnomalibDataset:
-        """Concatenate this dataset with another dataset."""
-        assert isinstance(other_dataset, self.__class__), "Cannot concatenate datasets that are not of the same type."
-        assert self.is_setup, "Cannot concatenate uninitialized datasets. Call setup first."
-        assert other_dataset.is_setup, "Cannot concatenate uninitialized datasets. Call setup first."
+    def __add__(self, other_dataset: "AnomalibDataset") -> "AnomalibDataset":
+        """Concatenate this dataset with another dataset.
+
+        Args:
+            other_dataset (AnomalibDataset): Dataset to concatenate with.
+
+        Returns:
+            AnomalibDataset: Concatenated dataset.
+        """
+        if not isinstance(other_dataset, self.__class__):
+            msg = "Cannot concatenate datasets that are not of the same type."
+            raise TypeError(msg)
         dataset = copy.deepcopy(self)
         dataset.samples = pd.concat([self.samples, other_dataset.samples], ignore_index=True)
         return dataset
-
-    def setup(self) -> None:
-        """Load data/metadata into memory."""
-        if not self.is_setup:
-            self._setup()
-        assert self.is_setup, "setup() should set self._samples"
-
-    @abstractmethod
-    def _setup(self) -> DataFrame:
-        """Set up the data module.
-
-        This method should return a dataframe that contains the information needed by the dataloader to load each of
-        the dataset items into memory.
-        The dataframe must at least contain the following columns:
-            split: the subset to which the dataset item is assigned.
-            image_path: path to file system location where the image is stored.
-            label_index: index of the anomaly label, typically 0 for "normal" and 1 for "anomalous".
-            mask_path (if task == "segmentation"): path to the ground truth masks (for the anomalous images only).
-
-        Example:
-        |---|-------------------|-----------|-------------|------------------|-------|
-        |   | image_path        | label     | label_index | mask_path        | split |
-        |---|-------------------|-----------|-------------|------------------|-------|
-        | 0 | path/to/image.png | anomalous | 1           | path/to/mask.png | train |
-        |---|-------------------|-----------|-------------|------------------|-------|
-        """
-        raise NotImplementedError
