@@ -3,60 +3,61 @@
 # Copyright (c) https://github.com/vislearn/FrEIA
 # SPDX-License-Identifier: MIT
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from __future__ import annotations
-
-import warnings
-from typing import Callable
+import logging
+from collections.abc import Callable
+from typing import Any
 
 import torch
-import torch.nn.functional as F
 from FrEIA.modules import InvertibleModule
 from scipy.stats import special_ortho_group
-from torch import Tensor, nn
+from torch import nn
+from torch.nn import functional as F  # noqa: N812
+
+logger = logging.getLogger(__name__)
 
 
-def _global_scale_sigmoid_activation(input: Tensor) -> Tensor:
+def _global_scale_sigmoid_activation(input_tensor: torch.Tensor) -> torch.Tensor:
     """Global scale sigmoid activation.
 
     Args:
-        input (Tensor): Input tensor
+        input_tensor (torch.Tensor): Input tensor
 
     Returns:
         Tensor: Sigmoid activation
     """
-    return 10 * torch.sigmoid(input - 2.0)
+    return 10 * torch.sigmoid(input_tensor - 2.0)
 
 
-def _global_scale_softplus_activation(input: Tensor) -> Tensor:
+def _global_scale_softplus_activation(input_tensor: torch.Tensor) -> torch.Tensor:
     """Global scale softplus activation.
 
     Args:
-        input (Tensor): Input tensor
+        input_tensor (torch.Tensor): Input tensor
 
     Returns:
         Tensor: Softplus activation
     """
     softplus = nn.Softplus(beta=0.5)
-    return 0.1 * softplus(input)
+    return 0.1 * softplus(input_tensor)
 
 
-def _global_scale_exp_activation(input: Tensor) -> Tensor:
+def _global_scale_exp_activation(input_tensor: torch.Tensor) -> torch.Tensor:
     """Global scale exponential activation.
 
     Args:
-        input (Tensor): Input tensor
+        input_tensor (torch.Tensor): Input tensor
 
     Returns:
         Tensor: Exponential activation
     """
-    return torch.exp(input)
+    return torch.exp(input_tensor)
 
 
 class AllInOneBlock(InvertibleModule):
-    """Module combining the most common operations in a normalizing flow or similar model.
+    r"""Module combining the most common operations in a normalizing flow or similar model.
 
     It combines affine coupling, permutation, and global affine transformation
     ('ActNorm'). It can also be used as GIN coupling block, perform learned
@@ -66,7 +67,7 @@ class AllInOneBlock(InvertibleModule):
 
     .. math::
 
-        y = V\\,R \\; \\Psi(s_\\mathrm{global}) \\odot \\mathrm{Coupling}\\Big(R^{-1} V^{-1} x\\Big)+ t_\\mathrm{global}
+        y = V R \; \Psi(s_\mathrm{global}) \odot \mathrm{Coupling}\Big(R^{-1} V^{-1} x\Big)+ t_\mathrm{global}
 
     - The inverse pre-permutation of x (i.e. :math:`R^{-1} V^{-1}`) is optional (see
       ``reverse_permutation`` below).
@@ -75,22 +76,40 @@ class AllInOneBlock(InvertibleModule):
       below).
     - For the coupling, the input is split into :math:`x_1, x_2` along
       the channel dimension. Then the output of the coupling operation is the
-      two halves :math:`u = \\mathrm{concat}(u_1, u_2)`.
+      two halves :math:`u = \mathrm{concat}(u_1, u_2)`.
 
       .. math::
 
-          u_1 &= x_1 \\odot \\exp \\Big( \\alpha \\; \\mathrm{tanh}\\big( s(x_2) \\big)\\Big) + t(x_2) \\\\
+          u_1 &= x_1 \odot \exp \Big( \alpha \; \mathrm{tanh}\big( s(x_2) \big)\Big) + t(x_2) \\
           u_2 &= x_2
 
-      Because :math:`\\mathrm{tanh}(s) \\in [-1, 1]`, this clamping mechanism prevents
-      exploding values in the exponential. The hyperparameter :math:`\\alpha` can be adjusted.
+      Because :math:`\mathrm{tanh}(s) \in [-1, 1]`, this clamping mechanism prevents
+      exploding values in the exponential. The hyperparameter :math:`\alpha` can be adjusted.
 
+    Args:
+        subnet_constructor: class or callable ``f``, called as ``f(channels_in, channels_out)`` and
+            should return a torch.nn.Module. Predicts coupling coefficients :math:`s, t`.
+        affine_clamping: clamp the output of the multiplicative coefficients before
+            exponentiation to +/- ``affine_clamping`` (see :math:`\alpha` above).
+        gin_block: Turn the block into a GIN block from Sorrenson et al, 2019.
+            Makes it so that the coupling operations as a whole is volume preserving.
+        global_affine_init: Initial value for the global affine scaling :math:`s_\mathrm{global}`.
+        global_affine_init: ``'SIGMOID'``, ``'SOFTPLUS'``, or ``'EXP'``. Defines the activation to be used
+            on the beta for the global affine scaling (:math:`\Psi` above).
+        permute_soft: bool, whether to sample the permutation matrix :math:`R` from :math:`SO(N)`,
+            or to use hard permutations instead. Note, ``permute_soft=True`` is very slow
+            when working with >512 dimensions.
+        learned_householder_permutation: Int, if >0, turn on the matrix :math:`V` above, that represents
+            multiple learned householder reflections. Slow if large number.
+            Dubious whether it actually helps network performance.
+        reverse_permutation: Reverse the permutation before the block, as introduced by Putzky
+            et al, 2019. Turns on the :math:`R^{-1} V^{-1}` pre-multiplication above.
     """
 
     def __init__(
         self,
-        dims_in,
-        dims_c=[],
+        dims_in: list[tuple[int]],
+        dims_c: list[tuple[int]] | None = None,
         subnet_constructor: Callable | None = None,
         affine_clamping: float = 2.0,
         gin_block: bool = False,
@@ -99,36 +118,9 @@ class AllInOneBlock(InvertibleModule):
         permute_soft: bool = False,
         learned_householder_permutation: int = 0,
         reverse_permutation: bool = False,
-    ):
-        """
-        Args:
-          subnet_constructor:
-            class or callable ``f``, called as ``f(channels_in, channels_out)`` and
-            should return a torch.nn.Module. Predicts coupling coefficients :math:`s, t`.
-          affine_clamping:
-            clamp the output of the multiplicative coefficients before
-            exponentiation to +/- ``affine_clamping`` (see :math:`\\alpha` above).
-          gin_block:
-            Turn the block into a GIN block from Sorrenson et al, 2019.
-            Makes it so that the coupling operations as a whole is volume preserving.
-          global_affine_init:
-            Initial value for the global affine scaling :math:`s_\mathrm{global}`.
-          global_affine_init:
-            ``'SIGMOID'``, ``'SOFTPLUS'``, or ``'EXP'``. Defines the activation to be used
-            on the beta for the global affine scaling (:math:`\\Psi` above).
-          permute_soft:
-            bool, whether to sample the permutation matrix :math:`R` from :math:`SO(N)`,
-            or to use hard permutations instead. Note, ``permute_soft=True`` is very slow
-            when working with >512 dimensions.
-          learned_householder_permutation:
-            Int, if >0, turn on the matrix :math:`V` above, that represents
-            multiple learned householder reflections. Slow if large number.
-            Dubious whether it actually helps network performance.
-          reverse_permutation:
-            Reverse the permutation before the block, as introduced by Putzky
-            et al, 2019. Turns on the :math:`R^{-1} V^{-1}` pre-multiplication above.
-        """
-
+    ) -> None:
+        if dims_c is None:
+            dims_c = []
         super().__init__(dims_in, dims_c)
 
         channels = dims_in[0][0]
@@ -141,9 +133,10 @@ class AllInOneBlock(InvertibleModule):
             self.conditional = False
             self.condition_channels = 0
         else:
-            assert tuple(dims_c[0][1:]) == tuple(
-                dims_in[0][1:]
-            ), f"Dimensions of input and condition don't agree: {dims_c} vs {dims_in}."
+            if tuple(dims_c[0][1:]) != tuple(dims_in[0][1:]):
+                msg = f"Dimensions of input and condition don't agree: {dims_c} vs {dims_in}."
+                raise ValueError(msg)
+
             self.conditional = True
             self.condition_channels = sum(dc[0] for dc in dims_c)
 
@@ -154,7 +147,8 @@ class AllInOneBlock(InvertibleModule):
         try:
             self.permute_function = {0: F.linear, 1: F.conv1d, 2: F.conv2d, 3: F.conv3d}[self.input_rank]
         except KeyError:
-            raise ValueError(f"Data is {1 + self.input_rank}D. Must be 1D-4D.")
+            msg = f"Data is {1 + self.input_rank}D. Must be 1D-4D."
+            raise ValueError(msg) from None
 
         self.in_channels = channels
         self.clamp = affine_clamping
@@ -163,16 +157,14 @@ class AllInOneBlock(InvertibleModule):
         self.householder = learned_householder_permutation
 
         if permute_soft and channels > 512:
-            warnings.warn(
-                (
-                    "Soft permutation will take a very long time to initialize "
-                    f"with {channels} feature channels. Consider using hard permutation instead."
-                )
+            msg = (
+                "Soft permutation will take a very long time to initialize "
+                f"with {channels} feature channels. Consider using hard permutation instead."
             )
+            logger.warning(msg)
 
         # global_scale is used as the initial value for the global affine scale
         # (pre-activation). It is computed such that
-        # global_scale_activation(global_scale) = global_affine_init
         # the 'magic numbers' (specifically for sigmoid) scale the activation to
         # a sensible range.
         if global_affine_type == "SIGMOID":
@@ -185,7 +177,8 @@ class AllInOneBlock(InvertibleModule):
             global_scale = torch.log(torch.tensor(global_affine_init))
             self.global_scale_activation = _global_scale_exp_activation
         else:
-            raise ValueError('Global affine activation must be "SIGMOID", "SOFTPLUS" or "EXP"')
+            message = 'Global affine activation must be "SIGMOID", "SOFTPLUS" or "EXP"'
+            raise ValueError(message)
 
         self.global_scale = nn.Parameter(torch.ones(1, self.in_channels, *([1] * self.input_rank)) * global_scale)
         self.global_offset = nn.Parameter(torch.zeros(1, self.in_channels, *([1] * self.input_rank)))
@@ -207,31 +200,42 @@ class AllInOneBlock(InvertibleModule):
             self.w_0 = nn.Parameter(torch.FloatTensor(w), requires_grad=False)
         else:
             self.w_perm = nn.Parameter(
-                torch.FloatTensor(w).view(channels, channels, *([1] * self.input_rank)), requires_grad=False
+                torch.FloatTensor(w).view(channels, channels, *([1] * self.input_rank)),
+                requires_grad=False,
             )
             self.w_perm_inv = nn.Parameter(
-                torch.FloatTensor(w.T).view(channels, channels, *([1] * self.input_rank)), requires_grad=False
+                torch.FloatTensor(w.T).view(channels, channels, *([1] * self.input_rank)),
+                requires_grad=False,
             )
 
         if subnet_constructor is None:
-            raise ValueError("Please supply a callable subnet_constructor" "function or object (see docstring)")
+            message = "Please supply a callable subnet_constructor function or object (see docstring)"
+            raise ValueError(message)
         self.subnet = subnet_constructor(self.splits[0] + self.condition_channels, 2 * self.splits[1])
         self.last_jac = None
 
-    def _construct_householder_permutation(self):
-        """Computes a permutation matrix from the reflection vectors that are
-        learned internally as nn.Parameters."""
+    def _construct_householder_permutation(self) -> torch.Tensor:
+        """Compute a permutation matrix from the reflection vectors that are learned internally as nn.Parameters."""
         w = self.w_0
         for vk in self.vk_householder:
             w = torch.mm(w, torch.eye(self.in_channels).to(w.device) - 2 * torch.ger(vk, vk) / torch.dot(vk, vk))
 
-        for i in range(self.input_rank):
+        for _ in range(self.input_rank):
             w = w.unsqueeze(-1)
         return w
 
-    def _permute(self, x, rev=False):
-        """Performs the permutation and scaling after the coupling operation.
-        Returns transformed outputs and the LogJacDet of the scaling operation."""
+    def _permute(self, x: torch.Tensor, rev: bool = False) -> tuple[Any, float | torch.Tensor]:
+        """Perform the permutation and scaling after the coupling operation.
+
+        Returns transformed outputs and the LogJacDet of the scaling operation.
+
+        Args:
+            x (torch.Tensor): Input tensor
+            rev (bool, optional): Reverse the permutation. Defaults to False.
+
+        Returns:
+            tuple[Any, float | torch.Tensor]: Transformed outputs and the LogJacDet of the scaling operation.
+        """
         if self.GIN:
             scale = 1.0
             perm_log_jac = 0.0
@@ -241,22 +245,26 @@ class AllInOneBlock(InvertibleModule):
 
         if rev:
             return ((self.permute_function(x, self.w_perm_inv) - self.global_offset) / scale, perm_log_jac)
-        else:
-            return (self.permute_function(x * scale + self.global_offset, self.w_perm), perm_log_jac)
 
-    def _pre_permute(self, x, rev=False):
-        """Permutes before the coupling block, only used if
-        reverse_permutation is set"""
+        return (self.permute_function(x * scale + self.global_offset, self.w_perm), perm_log_jac)
+
+    def _pre_permute(self, x: torch.Tensor, rev: bool = False) -> torch.Tensor:
+        """Permute before the coupling block.
+
+        It is only used if reverse_permutation is set.
+        """
         if rev:
             return self.permute_function(x, self.w_perm)
-        else:
-            return self.permute_function(x, self.w_perm_inv)
 
-    def _affine(self, x, a, rev=False):
-        """Given the passive half, and the pre-activation outputs of the
+        return self.permute_function(x, self.w_perm_inv)
+
+    def _affine(self, x: torch.Tensor, a: torch.Tensor, rev: bool = False) -> tuple[Any, torch.Tensor]:
+        """Perform affine coupling operation.
+
+        Given the passive half, and the pre-activation outputs of the
         coupling subnetwork, perform the affine coupling operation.
-        Returns both the transformed inputs and the LogJacDet."""
-
+        Returns both the transformed inputs and the LogJacDet.
+        """
         # the entire coupling coefficient tensor is scaled down by a
         # factor of ten for stability and easier initialization.
         a *= 0.1
@@ -268,11 +276,22 @@ class AllInOneBlock(InvertibleModule):
 
         if not rev:
             return (x * torch.exp(sub_jac) + a[:, ch:], torch.sum(sub_jac, dim=self.sum_dims))
-        else:
-            return ((x - a[:, ch:]) * torch.exp(-sub_jac), -torch.sum(sub_jac, dim=self.sum_dims))
 
-    def forward(self, x, c=[], rev=False, jac=True):
-        """See base class docstring"""
+        return ((x - a[:, ch:]) * torch.exp(-sub_jac), -torch.sum(sub_jac, dim=self.sum_dims))
+
+    def forward(
+        self,
+        x: torch.Tensor,
+        c: list | None = None,
+        rev: bool = False,
+        jac: bool = True,
+    ) -> tuple[tuple[torch.Tensor], torch.Tensor]:
+        """See base class docstring."""
+        del jac  # Unused argument.
+
+        if c is None:
+            c = []
+
         if self.householder:
             self.w_perm = self._construct_householder_permutation()
             if rev or self.reverse_pre_permute:
@@ -286,10 +305,7 @@ class AllInOneBlock(InvertibleModule):
 
         x1, x2 = torch.split(x[0], self.splits, dim=1)
 
-        if self.conditional:
-            x1c = torch.cat([x1, *c], 1)
-        else:
-            x1c = x1
+        x1c = torch.cat([x1, *c], 1) if self.conditional else x1
 
         if not rev:
             a1 = self.subnet(x1c)
@@ -314,5 +330,14 @@ class AllInOneBlock(InvertibleModule):
 
         return (x_out,), log_jac_det
 
-    def output_dims(self, input_dims):
+    @staticmethod
+    def output_dims(input_dims: list[tuple[int]]) -> list[tuple[int]]:
+        """Output dimensions of the layer.
+
+        Args:
+            input_dims (list[tuple[int]]): Input dimensions.
+
+        Returns:
+            list[tuple[int]]: Output dimensions.
+        """
         return input_dims
