@@ -1,243 +1,84 @@
 """Anomalib Datasets."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from __future__ import annotations
-
+import importlib
 import logging
+from enum import Enum
+from itertools import chain
 
 from omegaconf import DictConfig, ListConfig
 
-from .avenue import Avenue
+from anomalib.utils.config import to_tuple
+
 from .base import AnomalibDataModule, AnomalibDataset
-from .btech import BTech
-from .folder import Folder
-from .folder_3d import Folder3D
-from .inference import InferenceDataset
-from .mvtec import MVTec
-from .mvtec_3d import MVTec3D
-from .shanghaitech import ShanghaiTech
-from .task_type import TaskType
-from .ucsd_ped import UCSDped
-from .visa import Visa
+from .depth import DepthDataFormat, Folder3D, MVTec3D
+from .image import BTech, Folder, ImageDataFormat, Kolektor, MVTec, Visa
+from .predict import PredictDataset
+from .utils import LabelName
+from .video import Avenue, ShanghaiTech, UCSDped, VideoDataFormat
 
 logger = logging.getLogger(__name__)
 
 
-def get_datamodule(config: DictConfig | ListConfig) -> AnomalibDataModule:
+DataFormat = Enum(  # type: ignore[misc]
+    "DataFormat",
+    {i.name: i.value for i in chain(DepthDataFormat, ImageDataFormat, VideoDataFormat)},
+)
+
+
+class UnknownDatamoduleError(ModuleNotFoundError): ...
+
+
+def get_datamodule(config: DictConfig | ListConfig | dict) -> AnomalibDataModule:
     """Get Anomaly Datamodule.
 
     Args:
-        config (DictConfig | ListConfig): Configuration of the anomaly model.
+        config (DictConfig | ListConfig | dict): Configuration of the anomaly model.
 
     Returns:
         PyTorch Lightning DataModule
     """
     logger.info("Loading the datamodule")
 
-    datamodule: AnomalibDataModule
+    if isinstance(config, dict):
+        config = DictConfig(config)
 
-    # convert center crop to tuple
-    center_crop = config.dataset.get("center_crop")
-    if center_crop is not None:
-        center_crop = (center_crop[0], center_crop[1])
+    try:
+        _config = config.data if "data" in config else config
+        if len(_config.class_path.split(".")) > 1:
+            module = importlib.import_module(".".join(_config.class_path.split(".")[:-1]))
+        else:
+            module = importlib.import_module("anomalib.data")
+    except ModuleNotFoundError as exception:
+        logger.exception(f"ModuleNotFoundError: {_config.class_path}")
+        raise UnknownDatamoduleError from exception
+    dataclass = getattr(module, _config.class_path.split(".")[-1])
+    init_args = {**_config.get("init_args", {})}  # get dict
+    if "image_size" in init_args:
+        init_args["image_size"] = to_tuple(init_args["image_size"])
 
-    if config.dataset.format.lower() == "mvtec":
-        datamodule = MVTec(
-            root=config.dataset.path,
-            category=config.dataset.category,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            task=config.dataset.task,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            test_split_mode=config.dataset.test_split_mode,
-            test_split_ratio=config.dataset.test_split_ratio,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "mvtec_3d":
-        datamodule = MVTec3D(
-            root=config.dataset.path,
-            category=config.dataset.category,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            task=config.dataset.task,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            test_split_mode=config.dataset.test_split_mode,
-            test_split_ratio=config.dataset.test_split_ratio,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "btech":
-        datamodule = BTech(
-            root=config.dataset.path,
-            category=config.dataset.category,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            task=config.dataset.task,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            test_split_mode=config.dataset.test_split_mode,
-            test_split_ratio=config.dataset.test_split_ratio,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "folder":
-        datamodule = Folder(
-            root=config.dataset.root,
-            normal_dir=config.dataset.normal_dir,
-            abnormal_dir=config.dataset.abnormal_dir,
-            task=config.dataset.task,
-            normal_test_dir=config.dataset.normal_test_dir,
-            mask_dir=config.dataset.mask_dir,
-            extensions=config.dataset.extensions,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            test_split_mode=config.dataset.test_split_mode,
-            test_split_ratio=config.dataset.test_split_ratio,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "folder_3d":
-        datamodule = Folder3D(
-            root=config.dataset.root,
-            normal_dir=config.dataset.normal_dir,
-            normal_depth_dir=config.dataset.normal_depth_dir,
-            abnormal_dir=config.dataset.abnormal_dir,
-            abnormal_depth_dir=config.dataset.abnormal_depth_dir,
-            task=config.dataset.task,
-            normal_test_dir=config.dataset.normal_test_dir,
-            normal_test_depth_dir=config.dataset.normal_test_depth_dir,
-            mask_dir=config.dataset.mask_dir,
-            extensions=config.dataset.extensions,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            test_split_mode=config.dataset.test_split_mode,
-            test_split_ratio=config.dataset.test_split_ratio,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "ucsdped":
-        datamodule = UCSDped(
-            root=config.dataset.path,
-            category=config.dataset.category,
-            task=config.dataset.task,
-            clip_length_in_frames=config.dataset.clip_length_in_frames,
-            frames_between_clips=config.dataset.frames_between_clips,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "avenue":
-        datamodule = Avenue(
-            root=config.dataset.path,
-            gt_dir=config.dataset.gt_dir,
-            task=config.dataset.task,
-            clip_length_in_frames=config.dataset.clip_length_in_frames,
-            frames_between_clips=config.dataset.frames_between_clips,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "visa":
-        datamodule = Visa(
-            root=config.dataset.path,
-            category=config.dataset.category,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            task=config.dataset.task,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            test_split_mode=config.dataset.test_split_mode,
-            test_split_ratio=config.dataset.test_split_ratio,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    elif config.dataset.format.lower() == "shanghaitech":
-        datamodule = ShanghaiTech(
-            root=config.dataset.path,
-            scene=config.dataset.scene,
-            task=config.dataset.task,
-            clip_length_in_frames=config.dataset.clip_length_in_frames,
-            frames_between_clips=config.dataset.frames_between_clips,
-            image_size=(config.dataset.image_size[0], config.dataset.image_size[1]),
-            center_crop=center_crop,
-            normalization=config.dataset.normalization,
-            transform_config_train=config.dataset.transform_config.train,
-            transform_config_eval=config.dataset.transform_config.eval,
-            train_batch_size=config.dataset.train_batch_size,
-            eval_batch_size=config.dataset.eval_batch_size,
-            num_workers=config.dataset.num_workers,
-            val_split_mode=config.dataset.val_split_mode,
-            val_split_ratio=config.dataset.val_split_ratio,
-        )
-    else:
-        raise ValueError(
-            "Unknown dataset! \n"
-            "If you use a custom dataset make sure you initialize it in"
-            "`get_datamodule` in `anomalib.data.__init__.py"
-        )
-
-    return datamodule
+    return dataclass(**init_args)
 
 
 __all__ = [
     "AnomalibDataset",
     "AnomalibDataModule",
+    "DepthDataFormat",
+    "ImageDataFormat",
+    "VideoDataFormat",
     "get_datamodule",
     "BTech",
     "Folder",
     "Folder3D",
-    "InferenceDataset",
+    "PredictDataset",
+    "Kolektor",
     "MVTec",
     "MVTec3D",
     "Avenue",
     "UCSDped",
-    "TaskType",
     "ShanghaiTech",
+    "Visa",
+    "LabelName",
 ]
