@@ -1,6 +1,6 @@
 """Helper utilities for data."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 from .augmenter import Augmenter
@@ -13,17 +13,18 @@ from .image import (
     get_image_height_and_width,
     read_depth_image,
     read_image,
+    read_mask,
 )
-from .path import _check_and_convert_path, _prepare_files_labels, _resolve_path
-from .split import (
-    Split,
-    TestSplitMode,
-    ValSplitMode,
-    concatenate_datasets,
-    random_split,
-    split_by_label,
+from .label import LabelName
+from .path import (
+    DirType,
+    _check_and_convert_path,
+    _prepare_files_labels,
+    resolve_path,
+    validate_and_resolve_path,
+    validate_path,
 )
-from .transform import InputNormalizationMethod, get_transforms
+from .split import Split, TestSplitMode, ValSplitMode, concatenate_datasets, random_split, split_by_label
 
 __all__ = [
     "generate_output_image_filename",
@@ -31,6 +32,7 @@ __all__ = [
     "get_image_height_and_width",
     "random_2d_perlin",
     "read_image",
+    "read_mask",
     "read_depth_image",
     "random_split",
     "split_by_label",
@@ -38,15 +40,17 @@ __all__ = [
     "Split",
     "ValSplitMode",
     "TestSplitMode",
+    "LabelName",
+    "DirType",
     "Augmenter",
     "masks_to_boxes",
     "boxes_to_masks",
     "boxes_to_anomaly_maps",
-    "get_transforms",
-    "InputNormalizationMethod",
     "download_and_extract",
     "DownloadInfo",
     "_check_and_convert_path",
     "_prepare_files_labels",
-    "_resolve_path",
+    "resolve_path",
+    "validate_path",
+    "validate_and_resolve_path",
 ]
