@@ -1,17 +1,9 @@
 """Functions for Inference and model deployment."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .export import ExportMode, export, get_metadata, get_model_metadata
+from .export import CompressionType, ExportType
 from .inferencers import Inferencer, OpenVINOInferencer, TorchInferencer
 
-__all__ = [
-    "ExportMode",
-    "Inferencer",
-    "OpenVINOInferencer",
-    "TorchInferencer",
-    "export",
-    "get_model_metadata",
-    "get_metadata",
-]
+__all__ = ["Inferencer", "OpenVINOInferencer", "TorchInferencer", "ExportType", "CompressionType"]
