@@ -1,167 +1,113 @@
 """Utilities for optimization and OpenVINO conversion."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from __future__ import annotations
-
-import json
-import subprocess  # nosec
+import logging
 from enum import Enum
-from pathlib import Path
-from typing import Any
 
-import numpy as np
 import torch
-from torch import Tensor
-from torch.types import Number
+from torch import nn
+from torchvision.transforms.v2 import CenterCrop, Compose, Resize, Transform
 
-from anomalib.data.task_type import TaskType
-from anomalib.models.components import AnomalyModule
+from anomalib.data.transforms import ExportableCenterCrop
+
+logger = logging.getLogger("anomalib")
 
 
-class ExportMode(str, Enum):
-    """Model export mode."""
+class ExportType(str, Enum):
+    """Model export type.
+
+    Examples:
+        >>> from anomalib.deploy import ExportType
+        >>> ExportType.ONNX
+        'onnx'
+        >>> ExportType.OPENVINO
+        'openvino'
+        >>> ExportType.TORCH
+        'torch'
+    """
 
     ONNX = "onnx"
     OPENVINO = "openvino"
     TORCH = "torch"
 
 
-def get_model_metadata(model: AnomalyModule) -> dict[str, Tensor]:
-    """Get meta data related to normalization from model.
+class CompressionType(str, Enum):
+    """Model compression type when exporting to OpenVINO.
+
+    Attributes:
+        FP16 (str): Weight compression (FP16). All weights are converted to FP16.
+        INT8 (str): Weight compression (INT8). All weights are quantized to INT8,
+            but are dequantized to floating point before inference.
+        INT8_PTQ (str): Full integer post-training quantization (INT8).
+            All weights and operations are quantized to INT8. Inference is done
+            in INT8 precision.
+        INT8_ACQ (str): Accuracy-control quantization (INT8). Weights and
+            operations are quantized to INT8, except those that would degrade
+            quality of the model more than is acceptable. Inference is done in
+            a mixed precision.
+
+    Examples:
+        >>> from anomalib.deploy import CompressionType
+        >>> CompressionType.INT8_PTQ
+        'int8_ptq'
+    """
+
+    FP16 = "fp16"
+    INT8 = "int8"
+    INT8_PTQ = "int8_ptq"
+    INT8_ACQ = "int8_acq"
+
+
+class InferenceModel(nn.Module):
+    """Inference model for export.
+
+    The InferenceModel is used to wrap the model and transform for exporting to torch and ONNX/OpenVINO.
 
     Args:
-        model (AnomalyModule): Anomaly model which contains metadata related to normalization.
-
-    Returns:
-        dict[str, Tensor]: Model metadata
+        model (nn.Module): Model to export.
+        transform (Transform): Input transform for the model.
+        disable_antialias (bool, optional): Disable antialiasing in the Resize transforms of the given transform. This
+            is needed for ONNX/OpenVINO export, as antialiasing is not supported in the ONNX opset.
     """
-    metadata = {}
-    cached_metadata: dict[str, Number | Tensor] = {
-        "image_threshold": model.image_threshold.cpu().value.item(),
-        "pixel_threshold": model.pixel_threshold.cpu().value.item(),
-    }
-    if hasattr(model, "normalization_metrics") and model.normalization_metrics.state_dict() is not None:
-        for key, value in model.normalization_metrics.state_dict().items():
-            cached_metadata[key] = value.cpu()
-    # Remove undefined values by copying in a new dict
-    for key, val in cached_metadata.items():
-        if not np.isinf(val).all():
-            metadata[key] = val
-    del cached_metadata
-    return metadata
 
+    def __init__(self, model: nn.Module, transform: Transform, disable_antialias: bool = False) -> None:
+        super().__init__()
+        self.model = model
+        self.transform = transform
+        self.convert_center_crop()
+        if disable_antialias:
+            self.disable_antialias()
 
-def get_metadata(task: TaskType, transform: dict[str, Any], model: AnomalyModule) -> dict[str, Any]:
-    """Get metadata for the exported model.
+    def forward(self, batch: torch.Tensor) -> torch.Tensor | tuple[torch.Tensor, torch.Tensor]:
+        """Transform the input batch and pass it through the model."""
+        batch = self.transform(batch)
+        return self.model(batch)
 
-    Args:
-        task (TaskType): Task type.
-        transform (dict[str, Any]): Transform used for the model.
-        model (AnomalyModule): Model to export.
-        export_mode (ExportMode): Mode to export the model. Torch, ONNX or OpenVINO.
+    def disable_antialias(self) -> None:
+        """Disable antialiasing in the Resize transforms of the given transform.
 
-    Returns:
-        dict[str, Any]: Metadata for the exported model.
-    """
-    data_metadata = {"task": task, "transform": transform}
-    model_metadata = get_model_metadata(model)
-    metadata = {**data_metadata, **model_metadata}
+        This is needed for ONNX/OpenVINO export, as antialiasing is not supported in the ONNX opset.
+        """
+        if isinstance(self.transform, Resize):
+            self.transform.antialias = False
+        if isinstance(self.transform, Compose):
+            for transform in self.transform.transforms:
+                if isinstance(transform, Resize):
+                    transform.antialias = False
 
-    # Convert torch tensors to python lists or values for json serialization.
-    for key, value in metadata.items():
-        if isinstance(value, Tensor):
-            metadata[key] = value.numpy().tolist()
+    def convert_center_crop(self) -> None:
+        """Convert CenterCrop to ExportableCenterCrop for ONNX export.
 
-    return metadata
-
-
-def export(
-    task: TaskType,
-    transform: dict[str, Any],
-    input_size: tuple[int, int],
-    model: AnomalyModule,
-    export_mode: ExportMode,
-    export_root: str | Path,
-) -> None:
-    """Export the model to onnx format and (optionally) convert to OpenVINO IR if export mode is set to OpenVINO.
-
-    Args:
-        task (TaskType): Task type.
-        transform (dict[str, Any]): Data transforms (augmentatiions) used for the model.
-        input_size (tuple[int, int]): Input size of the model.
-        model (AnomalyModule): Anomaly model to export.
-        export_mode (ExportMode): Mode to export the model. Torch, ONNX or OpenVINO.
-        export_root (str | Path): Path to exported Torch, ONNX or OpenVINO IR.
-    """
-    # Create export directory.
-    export_path = Path(export_root) / "weights" / export_mode.value
-    export_path.mkdir(parents=True, exist_ok=True)
-
-    # Get metadata.
-    metadata = get_metadata(task, transform, model)
-
-    if export_mode == ExportMode.TORCH:
-        export_to_torch(model, metadata, export_path)
-
-    elif export_mode in (ExportMode.ONNX, ExportMode.OPENVINO):
-        # Write metadata to json file. The file is written in the same directory as the target model.
-        with (Path(export_path) / "metadata.json").open("w", encoding="utf-8") as metadata_file:
-            json.dump(metadata, metadata_file, ensure_ascii=False, indent=4)
-
-        # Export model to onnx and convert to OpenVINO IR if export mode is set to OpenVINO.
-        onnx_path = export_to_onnx(model, input_size, export_path)
-        if export_mode == ExportMode.OPENVINO:
-            export_to_openvino(export_path, onnx_path)
-
-    else:
-        raise ValueError(f"Unknown export mode {export_mode}")
-
-
-def export_to_torch(model: AnomalyModule, metadata: dict[str, Any], export_path: Path) -> None:
-    """Export AnomalibModel to torch.
-
-    Args:
-        model (AnomalyModule): Model to export.
-        export_path (Path): Path to the folder storing the exported model.
-    """
-    torch.save(
-        obj={"model": model.model, "metadata": metadata},
-        f=export_path / "model.pt",
-    )
-
-
-def export_to_onnx(model: AnomalyModule, input_size: tuple[int, int], export_path: Path) -> Path:
-    """Export model to onnx.
-
-    Args:
-        model (AnomalyModule): Model to export.
-        input_size (list[int] | tuple[int, int]): Image size used as the input for onnx converter.
-        export_path (Path): Path to the root folder of the exported model.
-
-    Returns:
-        Path: Path to the exported onnx model.
-    """
-    onnx_path = export_path / "model.onnx"
-    torch.onnx.export(
-        model.model,
-        torch.zeros((1, 3, *input_size)).to(model.device),
-        onnx_path,
-        opset_version=11,
-        input_names=["input"],
-        output_names=["output"],
-    )
-
-    return onnx_path
-
-
-def export_to_openvino(export_path: str | Path, onnx_path: Path) -> None:
-    """Convert onnx model to OpenVINO IR.
-
-    Args:
-        export_path (str | Path): Path to the root folder of the exported model.
-        onnx_path (Path): Path to the exported onnx model.
-    """
-    optimize_command = ["mo", "--input_model", str(onnx_path), "--output_dir", str(export_path)]
-    subprocess.run(optimize_command, check=True)  # nosec
+        The original CenterCrop transform is not supported in ONNX export. This method replaces the CenterCrop to
+        ExportableCenterCrop, which is supported in ONNX export. For more details, see the implementation of
+        ExportableCenterCrop.
+        """
+        if isinstance(self.transform, CenterCrop):
+            self.transform = ExportableCenterCrop(size=self.transform.size)
+        elif isinstance(self.transform, Compose):
+            transforms = self.transform.transforms
+            for index in range(len(transforms)):
+                if isinstance(transforms[index], CenterCrop):
+                    transforms[index] = ExportableCenterCrop(size=transforms[index].size)
