@@ -1,23 +1,20 @@
 """Components used within the models."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .base import AnomalyModule, DynamicBufferModule
+from .base import AnomalyModule, BufferListMixin, DynamicBufferMixin, MemoryBankMixin
 from .dimensionality_reduction import PCA, SparseRandomProjection
-from .feature_extractors import (
-    FeatureExtractor,
-    TimmFeatureExtractor,
-    TorchFXFeatureExtractor,
-)
+from .feature_extractors import TimmFeatureExtractor, TorchFXFeatureExtractor
 from .filters import GaussianBlur2d
 from .sampling import KCenterGreedy
 from .stats import GaussianKDE, MultiVariateGaussian
 
 __all__ = [
     "AnomalyModule",
-    "DynamicBufferModule",
-    "FeatureExtractor",
+    "BufferListMixin",
+    "DynamicBufferMixin",
+    "MemoryBankMixin",
     "GaussianKDE",
     "GaussianBlur2d",
     "KCenterGreedy",
