@@ -1,9 +1,11 @@
 """Base classes for all anomaly components."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 from .anomaly_module import AnomalyModule
-from .dynamic_module import DynamicBufferModule
+from .buffer_list import BufferListMixin
+from .dynamic_buffer import DynamicBufferMixin
+from .memory_bank_module import MemoryBankMixin
 
-__all__ = ["AnomalyModule", "DynamicBufferModule"]
+__all__ = ["AnomalyModule", "BufferListMixin", "DynamicBufferMixin", "MemoryBankMixin"]
