@@ -1,16 +1,14 @@
 """Kernel Density Estimation Classifier."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-
-from __future__ import annotations
 
 import logging
 import random
 from enum import Enum
 
 import torch
-from torch import Tensor, nn
+from torch import nn
 
 from anomalib.models.components import PCA, GaussianKDE
 
@@ -53,18 +51,21 @@ class KDEClassifier(nn.Module):
         self.register_buffer("max_length", torch.empty([]))
         self.max_length = torch.empty([])
 
-    def pre_process(self, feature_stack: Tensor, max_length: Tensor | None = None) -> tuple[Tensor, Tensor]:
+    def pre_process(
+        self,
+        feature_stack: torch.Tensor,
+        max_length: torch.Tensor | None = None,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
         """Pre-process the CNN features.
 
         Args:
-          feature_stack (Tensor): Features extracted from CNN
+          feature_stack (torch.Tensor): Features extracted from CNN
           max_length (Tensor | None): Used to unit normalize the feature_stack vector. If ``max_len`` is not
             provided, the length is calculated from the ``feature_stack``. Defaults to None.
 
         Returns:
             (Tuple): Stacked features and length
         """
-
         if max_length is None:
             max_length = torch.max(torch.linalg.norm(feature_stack, ord=2, dim=1))
 
@@ -73,19 +74,19 @@ class KDEClassifier(nn.Module):
         elif self.feature_scaling_method == FeatureScalingMethod.SCALE:
             feature_stack /= max_length
         else:
-            raise RuntimeError("Unknown pre-processing mode. Available modes are: Normalized and Scale.")
+            msg = "Unknown pre-processing mode. Available modes are: Normalized and Scale."
+            raise RuntimeError(msg)
         return feature_stack, max_length
 
-    def fit(self, embeddings: Tensor) -> bool:
+    def fit(self, embeddings: torch.Tensor) -> bool:
         """Fit a kde model to embeddings.
 
         Args:
-            embeddings (Tensor): Input embeddings to fit the model.
+            embeddings (torch.Tensor): Input embeddings to fit the model.
 
         Returns:
             Boolean confirming whether the training is successful.
         """
-
         if embeddings.shape[0] < self.n_pca_components:
             logger.info("Not enough features to commit. Not making a model.")
             return False
@@ -104,20 +105,19 @@ class KDEClassifier(nn.Module):
 
         return True
 
-    def compute_kde_scores(self, features: Tensor, as_log_likelihood: bool | None = False) -> Tensor:
+    def compute_kde_scores(self, features: torch.Tensor, as_log_likelihood: bool | None = False) -> torch.Tensor:
         """Compute the KDE scores.
 
         The scores calculated from the KDE model are converted to densities. If `as_log_likelihood` is set to true then
             the log of the scores are calculated.
 
         Args:
-            features (Tensor): Features to which the PCA model is fit.
+            features (torch.Tensor): Features to which the PCA model is fit.
             as_log_likelihood (bool | None, optional): If true, gets log likelihood scores. Defaults to False.
 
         Returns:
-            (Tensor): Score
+            (torch.Tensor): Score
         """
-
         features = self.pca_model.transform(features)
         features, _ = self.pre_process(features, self.max_length)
         # Scores are always assumed to be passed as a density
@@ -132,32 +132,29 @@ class KDEClassifier(nn.Module):
         return kde_scores
 
     @staticmethod
-    def compute_probabilities(scores: Tensor) -> Tensor:
-        """Converts density scores to anomaly probabilities (see https://www.desmos.com/calculator/ifju7eesg7).
+    def compute_probabilities(scores: torch.Tensor) -> torch.Tensor:
+        """Convert density scores to anomaly probabilities (see https://www.desmos.com/calculator/ifju7eesg7).
 
         Args:
-          scores (Tensor): density of an image.
+          scores (torch.Tensor): density of an image.
 
         Returns:
           probability that image with {density} is anomalous
         """
         return 1 / (1 + torch.exp(0.05 * (scores - 12)))
 
-    def predict(self, features: Tensor) -> Tensor:
+    def predict(self, features: torch.Tensor) -> torch.Tensor:
         """Predicts the probability that the features belong to the anomalous class.
 
         Args:
-          features (Tensor): Feature from which the output probabilities are detected.
+          features (torch.Tensor): Feature from which the output probabilities are detected.
 
         Returns:
           Detection probabilities
         """
-
         scores = self.compute_kde_scores(features, as_log_likelihood=True)
-        probabilities = self.compute_probabilities(scores)
+        return self.compute_probabilities(scores)
 
-        return probabilities
-
-    def forward(self, features: Tensor) -> Tensor:
+    def forward(self, features: torch.Tensor) -> torch.Tensor:
         """Make predictions on extracted features."""
         return self.predict(features)
