@@ -1,16 +1,15 @@
 """Feature extractors."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .timm import FeatureExtractor, TimmFeatureExtractor
+from .timm import TimmFeatureExtractor
 from .torchfx import BackboneParams, TorchFXFeatureExtractor
 from .utils import dryrun_find_featuremap_dims
 
 __all__ = [
     "BackboneParams",
     "dryrun_find_featuremap_dims",
-    "FeatureExtractor",
     "TimmFeatureExtractor",
     "TorchFXFeatureExtractor",
 ]
