@@ -3,17 +3,15 @@
 This script extracts features from a CNN network
 """
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from __future__ import annotations
-
 import logging
-import warnings
+from collections.abc import Sequence
 
 import timm
 import torch
-from torch import Tensor, nn
+from torch import nn
 
 logger = logging.getLogger(__name__)
 
@@ -30,28 +28,48 @@ class TimmFeatureExtractor(nn.Module):
             computation is required.
 
     Example:
-        >>> import torch
-        >>> from anomalib.models.components.feature_extractors import TimmFeatureExtractor
+        .. code-block:: python
 
-        >>> model = TimmFeatureExtractor(model="resnet18", layers=['layer1', 'layer2', 'layer3'])
-        >>> input = torch.rand((32, 3, 256, 256))
-        >>> features = model(input)
+            import torch
+            from anomalib.models.components.feature_extractors import TimmFeatureExtractor
 
-        >>> [layer for layer in features.keys()]
-            ['layer1', 'layer2', 'layer3']
-        >>> [feature.shape for feature in features.values()]
-            [torch.Size([32, 64, 64, 64]), torch.Size([32, 128, 32, 32]), torch.Size([32, 256, 16, 16])]
+            model = TimmFeatureExtractor(model="resnet18", layers=['layer1', 'layer2', 'layer3'])
+            input = torch.rand((32, 3, 256, 256))
+            features = model(input)
+
+            print([layer for layer in features.keys()])
+            # Output: ['layer1', 'layer2', 'layer3']
+
+            print([feature.shape for feature in features.values()]()
+            # Output: [torch.Size([32, 64, 64, 64]), torch.Size([32, 128, 32, 32]), torch.Size([32, 256, 16, 16])]
     """
 
-    def __init__(self, backbone: str, layers: list[str], pre_trained: bool = True, requires_grad: bool = False):
+    def __init__(
+        self,
+        backbone: str,
+        layers: Sequence[str],
+        pre_trained: bool = True,
+        requires_grad: bool = False,
+    ) -> None:
         super().__init__()
+
+        # Extract backbone-name and weight-URI from the backbone string.
+        if "__AT__" in backbone:
+            backbone, uri = backbone.split("__AT__")
+            pretrained_cfg = timm.models.registry.get_pretrained_cfg(backbone)
+            # Override pretrained_cfg["url"] to use different pretrained weights.
+            pretrained_cfg["url"] = uri
+        else:
+            pretrained_cfg = None
+
         self.backbone = backbone
-        self.layers = layers
+        self.layers = list(layers)
         self.idx = self._map_layer_to_idx()
         self.requires_grad = requires_grad
         self.feature_extractor = timm.create_model(
             backbone,
             pretrained=pre_trained,
+            pretrained_cfg=pretrained_cfg,
             features_only=True,
             exportable=True,
             out_indices=self.idx,
@@ -59,59 +77,53 @@ class TimmFeatureExtractor(nn.Module):
         self.out_dims = self.feature_extractor.feature_info.channels()
         self._features = {layer: torch.empty(0) for layer in self.layers}
 
-    def _map_layer_to_idx(self, offset: int = 3) -> list[int]:
-        """Maps set of layer names to indices of model.
-
-        Args:
-            offset (int) `timm` ignores the first few layers when indexing please update offset based on need
+    def _map_layer_to_idx(self) -> list[int]:
+        """Map set of layer names to indices of model.
 
         Returns:
-            Feature map extracted from the CNN
+            list[int]: Feature map extracted from the CNN.
         """
         idx = []
-        features = timm.create_model(
+        model = timm.create_model(
             self.backbone,
             pretrained=False,
-            features_only=False,
+            features_only=True,
             exportable=True,
         )
-        for i in self.layers:
+        # model.feature_info.info returns list of dicts containing info, inside which "module" contains layer name
+        layer_names = [info["module"] for info in model.feature_info.info]
+        for layer in self.layers:
             try:
-                idx.append(list(dict(features.named_children()).keys()).index(i) - offset)
-            except ValueError:
-                warnings.warn(f"Layer {i} not found in model {self.backbone}")
+                idx.append(layer_names.index(layer))
+            except ValueError:  # noqa: PERF203
+                msg = f"Layer {layer} not found in model {self.backbone}. Available layers: {layer_names}"
+                logger.warning(msg)
                 # Remove unfound key from layer dict
-                self.layers.remove(i)
+                self.layers.remove(layer)
 
         return idx
 
-    def forward(self, inputs: Tensor) -> dict[str, Tensor]:
+    def forward(self, inputs: torch.Tensor) -> dict[str, torch.Tensor]:
         """Forward-pass input tensor into the CNN.
 
         Args:
-            inputs (Tensor): Input tensor
+            inputs (torch.Tensor): Input tensor
 
         Returns:
             Feature map extracted from the CNN
+
+        Example:
+            .. code-block:: python
+
+                model = TimmFeatureExtractor(model="resnet50", layers=['layer3'])
+                input = torch.rand((32, 3, 256, 256))
+                features = model.forward(input)
+
         """
         if self.requires_grad:
-            features = dict(zip(self.layers, self.feature_extractor(inputs)))
+            features = dict(zip(self.layers, self.feature_extractor(inputs), strict=True))
         else:
             self.feature_extractor.eval()
             with torch.no_grad():
-                features = dict(zip(self.layers, self.feature_extractor(inputs)))
+                features = dict(zip(self.layers, self.feature_extractor(inputs), strict=True))
         return features
-
-
-class FeatureExtractor(TimmFeatureExtractor):
-    """Compatibility wrapper for the old FeatureExtractor class.
-
-    See :class:`anomalib.models.components.feature_extractors.timm.TimmFeatureExtractor` for more details.
-    """
-
-    def __init__(self, *args, **kwargs):
-        logger.warning(
-            "FeatureExtractor is deprecated. Use TimmFeatureExtractor instead."
-            " Both FeatureExtractor and TimmFeatureExtractor will be removed in a future release."
-        )
-        super().__init__(*args, **kwargs)
