@@ -1,15 +1,16 @@
 """Utility functions to manipulate feature extractors."""
 
-from __future__ import annotations
+# Copyright (C) 2022-2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
 
 import torch
 from torch.fx.graph_module import GraphModule
 
-from .timm import FeatureExtractor
+from .timm import TimmFeatureExtractor
 
 
 def dryrun_find_featuremap_dims(
-    feature_extractor: FeatureExtractor | GraphModule,
+    feature_extractor: TimmFeatureExtractor | GraphModule,
     input_size: tuple[int, int],
     layers: list[str],
 ) -> dict[str, dict[str, int | tuple[int, int]]]:
@@ -19,8 +20,8 @@ def dryrun_find_featuremap_dims(
         tuple[int, int]: maping of `layer -> dimensions dict`
             Each `dimension dict` has two keys: `num_features` (int) and `resolution`(tuple[int, int]).
     """
-
-    dryrun_input = torch.empty(1, 3, *input_size)
+    device = next(feature_extractor.parameters()).device
+    dryrun_input = torch.empty(1, 3, *input_size).to(device)
     dryrun_features = feature_extractor(dryrun_input)
     return {
         layer: {"num_features": dryrun_features[layer].shape[1], "resolution": dryrun_features[layer].shape[2:]}
