@@ -1,44 +1,41 @@
 """Gaussian Kernel Density Estimation."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-
-from __future__ import annotations
 
 import math
 
 import torch
-from torch import Tensor
 
-from anomalib.models.components.base import DynamicBufferModule
+from anomalib.models.components.base import DynamicBufferMixin
 
 
-class GaussianKDE(DynamicBufferModule):
+class GaussianKDE(DynamicBufferMixin):
     """Gaussian Kernel Density Estimation.
 
     Args:
         dataset (Tensor | None, optional): Dataset on which to fit the KDE model. Defaults to None.
     """
 
-    def __init__(self, dataset: Tensor | None = None):
+    def __init__(self, dataset: torch.Tensor | None = None) -> None:
         super().__init__()
 
         if dataset is not None:
             self.fit(dataset)
 
-        self.register_buffer("bw_transform", Tensor())
-        self.register_buffer("dataset", Tensor())
-        self.register_buffer("norm", Tensor())
+        self.register_buffer("bw_transform", torch.Tensor())
+        self.register_buffer("dataset", torch.Tensor())
+        self.register_buffer("norm", torch.Tensor())
 
-        self.bw_transform = Tensor()
-        self.dataset = Tensor()
-        self.norm = Tensor()
+        self.bw_transform = torch.Tensor()
+        self.dataset = torch.Tensor()
+        self.norm = torch.Tensor()
 
-    def forward(self, features: Tensor) -> Tensor:
+    def forward(self, features: torch.Tensor) -> torch.Tensor:
         """Get the KDE estimates from the feature map.
 
         Args:
-          features (Tensor): Feature map extracted from the CNN
+          features (torch.Tensor): Feature map extracted from the CNN
 
         Returns: KDE Estimates
         """
@@ -52,11 +49,11 @@ class GaussianKDE(DynamicBufferModule):
 
         return estimate
 
-    def fit(self, dataset: Tensor) -> None:
+    def fit(self, dataset: torch.Tensor) -> None:
         """Fit a KDE model to the input dataset.
 
         Args:
-          dataset (Tensor): Input dataset.
+          dataset (torch.Tensor): Input dataset.
 
         Returns:
             None
@@ -74,7 +71,6 @@ class GaussianKDE(DynamicBufferModule):
         bw_transform = torch.linalg.cholesky(inv_cov)
         dataset = torch.matmul(dataset, bw_transform)
 
-        #
         norm = torch.prod(torch.diag(bw_transform))
         norm *= math.pow((2 * math.pi), (-dimension / 2))
 
@@ -83,16 +79,15 @@ class GaussianKDE(DynamicBufferModule):
         self.norm = norm
 
     @staticmethod
-    def cov(tensor: Tensor) -> Tensor:
+    def cov(tensor: torch.Tensor) -> torch.Tensor:
         """Calculate the unbiased covariance matrix.
 
         Args:
-            tensor (Tensor): Input tensor from which covariance matrix is computed.
+            tensor (torch.Tensor): Input tensor from which covariance matrix is computed.
 
         Returns:
             Output covariance matrix.
         """
         mean = torch.mean(tensor, dim=1)
         tensor -= mean[:, None]
-        cov = torch.matmul(tensor, tensor.T) / (tensor.size(1) - 1)
-        return cov
+        return torch.matmul(tensor, tensor.T) / (tensor.size(1) - 1)
