@@ -7,14 +7,13 @@ import cv2
 import numpy as np
 import torch
 from kornia.contrib import connected_components
-from torch import Tensor
 
 
-def connected_components_gpu(image: Tensor, num_iterations: int = 1000) -> Tensor:
+def connected_components_gpu(image: torch.Tensor, num_iterations: int = 1000) -> torch.Tensor:
     """Perform connected component labeling on GPU and remap the labels from 0 to N.
 
     Args:
-        image (Tensor): Binary input image from which we want to extract connected components (Bx1xHxW)
+        image (torch.Tensor): Binary input image from which we want to extract connected components (Bx1xHxW)
         num_iterations (int): Number of iterations used in the connected component computation.
 
     Returns:
@@ -30,19 +29,19 @@ def connected_components_gpu(image: Tensor, num_iterations: int = 1000) -> Tenso
     return components.int()
 
 
-def connected_components_cpu(image: Tensor) -> Tensor:
-    """Connected component labeling on CPU.
+def connected_components_cpu(image: torch.Tensor) -> torch.Tensor:
+    """Perform connected component labeling on CPU.
 
     Args:
-        image (Tensor): Binary input data from which we want to extract connected components (Bx1xHxW)
+        image (torch.Tensor): Binary input data from which we want to extract connected components (Bx1xHxW)
 
     Returns:
         Tensor: Components labeled from 0 to N.
     """
     components = torch.zeros_like(image)
     label_idx = 1
-    for i, mask in enumerate(image):
-        mask = mask.squeeze().numpy().astype(np.uint8)
+    for i, msk in enumerate(image):
+        mask = msk.squeeze().cpu().numpy().astype(np.uint8)
         _, comps = cv2.connectedComponents(mask)
         # remap component values to make sure every component has a unique value when outputs are concatenated
         for label in np.unique(comps)[1:]:
