@@ -1,4 +1,4 @@
-"""Tests."""
+"""Anomalib Test Suite."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
