@@ -1,87 +1,64 @@
 """Inference Entrypoint script."""
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from argparse import ArgumentParser, Namespace
-from pathlib import Path
-
-from pytorch_lightning import Trainer
+from jsonargparse import ActionConfigFile, Namespace
+from lightning.pytorch.callbacks import Callback
+from lightning.pytorch.cli import LightningArgumentParser
 from torch.utils.data import DataLoader
 
-from anomalib.config import get_configurable_parameters
-from anomalib.data.inference import InferenceDataset
-from anomalib.data.utils import InputNormalizationMethod, get_transforms
-from anomalib.models import get_model
-from anomalib.utils.callbacks import get_callbacks
+from anomalib.data import PredictDataset
+from anomalib.engine import Engine
+from anomalib.models import AnomalyModule, get_model
 
 
-def get_args() -> Namespace:
-    """Get command line arguments.
+def get_parser() -> LightningArgumentParser:
+    """Get parser.
 
     Returns:
-        Namespace: List of arguments.
+        LightningArgumentParser: The parser object.
     """
-    parser = ArgumentParser()
-    parser.add_argument("--config", type=Path, required=True, help="Path to a config file")
-    parser.add_argument("--weights", type=Path, required=True, help="Path to model weights")
-    parser.add_argument("--input", type=Path, required=True, help="Path to image(s) to infer.")
+    parser = LightningArgumentParser(description="Inference on Anomaly models in Lightning format.")
+    parser.add_lightning_class_args(AnomalyModule, "model", subclass_mode=True)
+    parser.add_lightning_class_args(Callback, "--callbacks", subclass_mode=True, required=False)
+    parser.add_argument("--ckpt_path", type=str, required=True, help="Path to model weights")
+    parser.add_class_arguments(PredictDataset, "--data", instantiate=False)
     parser.add_argument("--output", type=str, required=False, help="Path to save the output image(s).")
-    parser.add_argument(
-        "--visualization_mode",
-        type=str,
-        required=False,
-        default="simple",
-        help="Visualization mode.",
-        choices=["full", "simple"],
-    )
     parser.add_argument(
         "--show",
         action="store_true",
         required=False,
         help="Show the visualized predictions on the screen.",
     )
-
-    args = parser.parse_args()
-    return args
-
-
-def infer():
-    """Run inference."""
-    args = get_args()
-    config = get_configurable_parameters(config_path=args.config)
-    config.trainer.resume_from_checkpoint = str(args.weights)
-    config.visualization.show_images = args.show
-    config.visualization.mode = args.visualization_mode
-    if args.output:  # overwrite save path
-        config.visualization.save_images = True
-        config.visualization.image_save_path = args.output
-    else:
-        config.visualization.save_images = False
-
-    # create model and trainer
-    model = get_model(config)
-    callbacks = get_callbacks(config)
-    trainer = Trainer(callbacks=callbacks, **config.trainer)
-
-    # get the transforms
-    transform_config = config.dataset.transform_config.eval if "transform_config" in config.dataset.keys() else None
-    image_size = (config.dataset.image_size[0], config.dataset.image_size[1])
-    center_crop = config.dataset.get("center_crop")
-    if center_crop is not None:
-        center_crop = tuple(center_crop)
-    normalization = InputNormalizationMethod(config.dataset.normalization)
-    transform = get_transforms(
-        config=transform_config, image_size=image_size, center_crop=center_crop, normalization=normalization
+    parser.add_argument(
+        "-c",
+        "--config",
+        action=ActionConfigFile,
+        help="Path to a configuration file in json or yaml format.",
     )
 
+    return parser
+
+
+def infer(args: Namespace) -> None:
+    """Run inference."""
+    callbacks = None if not hasattr(args, "callbacks") else args.callbacks
+    engine = Engine(
+        default_root_dir=args.output,
+        callbacks=callbacks,
+        devices=1,
+    )
+    model = get_model(args.model)
+
     # create the dataset
-    dataset = InferenceDataset(args.input, image_size=tuple(config.dataset.image_size), transform=transform)
+    dataset = PredictDataset(**args.data)
     dataloader = DataLoader(dataset)
 
-    # generate predictions
-    trainer.predict(model=model, dataloaders=[dataloader])
+    engine.predict(model=model, dataloaders=[dataloader], ckpt_path=args.ckpt_path)
 
 
 if __name__ == "__main__":
-    infer()
+    parser = get_parser()
+    args = parser.parse_args()
+    infer(args)
