@@ -1,33 +1,31 @@
-"""Anomalib Inferencer Script.
+"""Anomalib OpenVINO Inferencer Script.
 
-This script performs inference by reading a model config file from
-command line, and show the visualization results.
+This script performs OpenVINO inference by reading a model from
+file system, and show the visualization results.
 """
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-import warnings
+import logging
 from argparse import ArgumentParser, Namespace
 from pathlib import Path
 
-from anomalib.data.utils import (
-    generate_output_image_filename,
-    get_image_filenames,
-    read_image,
-)
+from anomalib.data.utils import generate_output_image_filename, get_image_filenames, read_image
+from anomalib.data.utils.image import save_image, show_image
 from anomalib.deploy import OpenVINOInferencer
-from anomalib.post_processing import Visualizer
+from anomalib.utils.visualization import ImageVisualizer
+
+logger = logging.getLogger(__name__)
 
 
-def get_args() -> Namespace:
-    """Get command line arguments.
+def get_parser() -> ArgumentParser:
+    """Get parser.
 
     Returns:
-        Namespace: List of arguments.
+        ArgumentParser: The parser object.
     """
     parser = ArgumentParser()
-    parser.add_argument("--config", type=Path, required=True, help="Path to a config file")
     parser.add_argument("--weights", type=Path, required=True, help="Path to model weights")
     parser.add_argument("--metadata", type=Path, required=True, help="Path to a JSON file containing the metadata.")
     parser.add_argument("--input", type=Path, required=True, help="Path to an image to infer.")
@@ -63,24 +61,20 @@ def get_args() -> Namespace:
         help="Show the visualized predictions on the screen.",
     )
 
-    args = parser.parse_args()
-
-    return args
+    return parser
 
 
-def infer() -> None:
+def infer(args: Namespace) -> None:
     """Infer predictions.
 
     Show/save the output if path is to an image. If the path is a directory, go over each image in the directory.
-    """
-    # Get the command line arguments, and config from the config.yaml file.
-    # This config file is also used for training and contains all the relevant
-    # information regarding the data, model, train and inference details.
-    args = get_args()
 
+    Args:
+        args (Namespace): The arguments from the command line.
+    """
     # Get the inferencer.
-    inferencer = OpenVINOInferencer(path=args.weights, metadata_path=args.metadata, device=args.device)
-    visualizer = Visualizer(mode=args.visualization_mode, task=args.task)
+    inferencer = OpenVINOInferencer(path=args.weights, metadata=args.metadata, device=args.device)
+    visualizer = ImageVisualizer(mode=args.visualization_mode, task=args.task)
 
     filenames = get_image_filenames(path=args.input)
     for filename in filenames:
@@ -89,18 +83,18 @@ def infer() -> None:
         output = visualizer.visualize_image(predictions)
 
         if args.output is None and args.show is False:
-            warnings.warn(
-                "Neither output path is provided nor show flag is set. Inferencer will run but return nothing."
-            )
+            msg = "Neither output path is provided nor show flag is set. Inferencer will run but return nothing."
+            logger.warning(msg)
 
         if args.output:
             file_path = generate_output_image_filename(input_path=filename, output_path=args.output)
-            visualizer.save(file_path=file_path, image=output)
+            save_image(filename=file_path, image=output)
 
         # Show the image in case the flag is set by the user.
         if args.show:
-            visualizer.show(title="Output Image", image=output)
+            show_image(title="Output Image", image=output)
 
 
 if __name__ == "__main__":
-    infer()
+    args = get_parser().parse_args()
+    infer(args)
